requirements.txt
altair==4.2.2
async-generator==1.10
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.1
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.0
mdurl==0.1.2
nltk==3.8.1
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
protobuf==3.20.3
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
rich==13.3.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.8.2
semver==2.13.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
streamlit==1.20.0
tenacity==8.2.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.0
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.3.0
wsproto==1.2.0
zipp==3.15.0